pyproject.toml
[tool.poetry]
name = "pyprac"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
bs4 = "^0.0.1"
aiofiles = "^0.6.0"
aiohttp = "^3.7.2"
requests = "^2.25.0"
pandas = "^1.1.4"
pyecharts = "^1.9.0"
pysnooper = "^0.4.2"
yamllint = "^1.25.0"
FuzzyWuzzy = "^0.18.0"
hypothesis = "^5.43.3"
pandasql = "^0.7.3"
eliot = "^1.12.0"
eliot-tree = "^19.0.1"
aiosql = "^3.2.0"
funcy = "^1.15"
numpy = "^1.20.1"
matplotlib = "^3.3.4"
APScheduler = "^3.7.0"
ExifRead = "^2.3.2"
fastapi = "^0.63.0"
databases = "^0.4.2"
aiosqlite = "^0.17.0"
SQLAlchemy = "1.3"
uvicorn = "^0.13.4"
Pillow = "^8.1.2"
opencv-python = "^4.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
